Extract #to_h and #to_json into a dedicated mixin
