+++ messages/ruby/lib/cucumber/messages/message.rb
@@ -1,10 +1,10 @@
 require 'json'
-require 'cucumber/messages/message/utils'
+require 'cucumber/messages/message/serialization'
 
 module Cucumber
   module Messages
     class Message
-      include Cucumber::Messages::Message::Utils
+      include Cucumber::Messages::Message::Serialization
 
       def self.from_json(json_string)
         message_hash = JSON.parse(json_string)
@@ -29,19 +29,6 @@ module Cucumber
           arguments = prepare_DTO_arguments(klass, value)
           klass.new(arguments)
         end
-      end
-
-      def to_h(camelize: false)
-        self.instance_variables.to_h do |variable_name|
-          key = variable_name[1..-1]
-          key = Cucumber::Messages::Message.camelize(key) if camelize
-
-          [key.to_sym, get_h_value(variable_name, camelize: camelize)]
-        end
-      end
-
-      def to_json
-        to_h(camelize: true).to_json
       end
 
       private
